reorganized code and added index files to facilitate importing the npm package
