+++ schemas/booleanAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/BooleanAnswer",
+  "definitions": {
+    "BooleanAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "boolean"
+        },
+        "answer": {
+          "type": "boolean"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
