+++ schemas/checkboxesAnswer.schema.json
@@ -1,0 +1,26 @@
+{
+  "$ref": "#/definitions/CheckboxesAnswer",
+  "definitions": {
+    "CheckboxesAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "checkBoxes"
+        },
+        "answer": {
+          "type": "array",
+          "items": {
+            "type": "string"
+          }
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
