+++ schemas/currencyAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/CurrencyAnswer",
+  "definitions": {
+    "CurrencyAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "currency"
+        },
+        "answer": {
+          "type": "number"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
