+++ schemas/currencyQuestion.schema.json
@@ -1,0 +1,45 @@
+{
+  "$ref": "#/definitions/CurrencyQuestion",
+  "definitions": {
+    "CurrencyQuestion": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "currency"
+        },
+        "meta": {
+          "type": "object",
+          "properties": {
+            "denomination": {
+              "type": "string"
+            }
+          },
+          "additionalProperties": false
+        },
+        "attributes": {
+          "type": "object",
+          "properties": {
+            "max": {
+              "type": "number"
+            },
+            "min": {
+              "type": "number"
+            },
+            "step": {
+              "type": "number"
+            }
+          },
+          "additionalProperties": false
+        }
+      },
+      "required": [
+        "type",
+        "meta",
+        "attributes"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
