+++ schemas/datePickerAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/DatePickerAnswer",
+  "definitions": {
+    "DatePickerAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "datePicker"
+        },
+        "answer": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
