+++ schemas/dateRangeAnswer.schema.json
@@ -1,0 +1,36 @@
+{
+  "$ref": "#/definitions/DateRangeAnswer",
+  "definitions": {
+    "DateRangeAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "dateRange"
+        },
+        "answer": {
+          "type": "object",
+          "properties": {
+            "start": {
+              "type": "string"
+            },
+            "end": {
+              "type": "string"
+            }
+          },
+          "required": [
+            "start",
+            "end"
+          ],
+          "additionalProperties": false
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
