+++ schemas/emailAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/EmailAnswer",
+  "definitions": {
+    "EmailAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "email"
+        },
+        "answer": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
