+++ schemas/filteredSearchAnswer.schema.json
@@ -1,0 +1,26 @@
+{
+  "$ref": "#/definitions/FilteredSearchAnswer",
+  "definitions": {
+    "FilteredSearchAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "filteredSearch"
+        },
+        "answer": {
+          "type": "array",
+          "items": {
+            "type": "string"
+          }
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
