+++ schemas/numberAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/NumberAnswer",
+  "definitions": {
+    "NumberAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "number"
+        },
+        "answer": {
+          "type": "number"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
