+++ schemas/radioButtonsAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/RadioButtonsAnswer",
+  "definitions": {
+    "RadioButtonsAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "radioButtons"
+        },
+        "answer": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
