+++ schemas/selectBoxAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/SelectBoxAnswer",
+  "definitions": {
+    "SelectBoxAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "selectBox"
+        },
+        "answer": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
