+++ schemas/textAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/TextAnswer",
+  "definitions": {
+    "TextAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "text"
+        },
+        "answer": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
