+++ schemas/textAreaAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/TextAreaAnswer",
+  "definitions": {
+    "TextAreaAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "textArea"
+        },
+        "answer": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
