+++ schemas/textAreaQuestion.schema.json
@@ -1,0 +1,48 @@
+{
+  "$ref": "#/definitions/TextAreaQuestion",
+  "definitions": {
+    "TextAreaQuestion": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "textArea"
+        },
+        "meta": {
+          "type": "object",
+          "properties": {
+            "asRichText": {
+              "type": "boolean"
+            }
+          },
+          "additionalProperties": false
+        },
+        "attributes": {
+          "type": "object",
+          "properties": {
+            "cols": {
+              "type": "number"
+            },
+            "maxLength": {
+              "type": "number"
+            },
+            "minLength": {
+              "type": "number"
+            },
+            "rows": {
+              "type": "number"
+            }
+          },
+          "additionalProperties": false
+        }
+      },
+      "required": [
+        "type",
+        "meta",
+        "attributes"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
