+++ schemas/typeaheadSearchAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/TypeaheadSearchAnswer",
+  "definitions": {
+    "TypeaheadSearchAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "typeaheadSearch"
+        },
+        "answer": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
