+++ schemas/urlAnswer.schema.json
@@ -1,0 +1,23 @@
+{
+  "$ref": "#/definitions/URLAnswer",
+  "definitions": {
+    "URLAnswer": {
+      "type": "object",
+      "properties": {
+        "type": {
+          "type": "string",
+          "const": "url"
+        },
+        "answer": {
+          "type": "string"
+        }
+      },
+      "required": [
+        "type",
+        "answer"
+      ],
+      "additionalProperties": false
+    }
+  },
+  "$schema": "http://json-schema.org/draft-07/schema#"
+}
